Auth using gates:

+++ app/Http/Controllers/QuestionsController.php
@@ -40,16 +40,27 @@ class QuestionsController extends Controller
     }
 
     public function edit(Question $question) {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Accesss denied.");
+        }
         return view("questions.edit", compact('question'));
     }
 
     public function update(AskQuestionRequest $request, Question $question) {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Accesss denied.");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated.');
     }
 
     public function destroy(Question $question) {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Accesss denied.");
+        }
+        
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deletec.');
